todo comments and a start on move choosing when not random
most of the work left is in chooseMove function, when not random. i set up an array to hold current reward values, and copied logic for checking if cells are full/if someone won from rand moves above. need to select cell with highest reward value and put a player's mark in it- python doesn't have built in argmax so this might be a pain